refactor: move logic to component
