+++ src/components/map/components/external-realtime-map/ExternalMapButton.tsx
@@ -2,12 +2,26 @@ import React from 'react';
 import {Button} from '@atb/components/button';
 import {StyleSheet} from '@atb/theme';
 import {Map} from '@atb/assets/svg/mono-icons/map';
+import {getTextForLanguage, useTranslation} from '@atb/translations';
+import {useFirestoreConfiguration} from '@atb/configuration';
+import {MapSelectionActionType} from '../../types';
 
 type ExternalMapButtonProps = {
-  onPress: () => void;
+  onMapClick: (sc: MapSelectionActionType) => void;
 };
-export const ExternalMapButton = ({onPress}: ExternalMapButtonProps) => {
+export const ExternalMapButton = ({onMapClick}: ExternalMapButtonProps) => {
   const style = useStyle();
+  const {language} = useTranslation();
+  const {configurableLinks} = useFirestoreConfiguration();
+
+  const externalRealtimeMapUrl = getTextForLanguage(
+    configurableLinks?.externalRealtimeMap,
+    language,
+  );
+
+  if (!externalRealtimeMapUrl) {
+    return null;
+  }
 
   return (
     <Button
@@ -16,7 +30,12 @@ export const ExternalMapButton = ({onPress}: ExternalMapButtonProps) => {
       compact={true}
       interactiveColor="interactive_2"
       accessibilityRole="button"
-      onPress={onPress}
+      onPress={() =>
+        onMapClick({
+          source: 'external-map-button',
+          url: externalRealtimeMapUrl,
+        })
+      }
       rightIcon={{svg: Map}}
       hasShadow={true}
     />
